+++ src/test/java/com/ontotext/refine/cli/CreateProjectTest.java
@@ -52,6 +52,24 @@ class CreateProjectTest extends BaseProcessTest {
   }
 
   @Test
+  @ExpectedSystemExit(ExitCode.USAGE)
+  void shouldFailForUnsupportedFormat() {
+    try {
+      URL resource = getClass().getClassLoader().getResource("Netherlands_restaurants.csv");
+
+      String uriArg = "-u " + responder.getUri();
+      commandExecutor().accept(args(resource.getPath(), "-n Restaurants", "-f tsv", uriArg));
+    } finally {
+      String[] errorsArray = consoleErrors().split(System.lineSeparator());
+      String lastLine = errorsArray[0];
+      assertEquals(
+          "Invalid value for option '--format': expected one of [CSV] (case-insensitive)"
+              + " but was 'tsv'",
+          lastLine);
+    }
+  }
+
+  @Test
   @ExpectedSystemExit(ExitCode.SOFTWARE)
   void shouldFailToGetCrsfToken() {
     try {
@@ -59,7 +77,7 @@ class CreateProjectTest extends BaseProcessTest {
       URL resource = getClass().getClassLoader().getResource("Netherlands_restaurants.csv");
 
       String uriArg = "-u " + responder.getUri();
-      commandExecutor().accept(args(resource.getPath(), "-name Restaurants", uriArg));
+      commandExecutor().accept(args(resource.getPath(), "-n Restaurants", uriArg));
     } finally {
       failCsrfRequest = false;
 
@@ -79,7 +97,7 @@ class CreateProjectTest extends BaseProcessTest {
       URL resource = getClass().getClassLoader().getResource("Netherlands_restaurants.csv");
 
       String uriArg = "-u " + responder.getUri();
-      commandExecutor().accept(args(resource.getPath(), "-name Restaurants", uriArg));
+      commandExecutor().accept(args(resource.getPath(), "-n Restaurants", "-f csv", uriArg));
     } finally {
       String errors = consoleErrors();
       assertTrue(errors.isEmpty(), "Expected no errors but there were: " + errors);